+++ EduVPN-multiOS/Coordinators/AppCoordinator.swift
@@ -37,6 +37,7 @@ class AppCoordinator: RootViewCoordinator {
     }()
     
     let persistentContainer = NSPersistentContainer(name: "EduVPN")
+    let notificationsService = NotificationsService()
     
     // MARK: - Properties
     
@@ -255,6 +256,7 @@ class AppCoordinator: RootViewCoordinator {
             .then { response -> Promise<CertificateModel> in response.mapResponse() }
             .map { model -> CertificateModel in
                 api.certificateModel = model
+                self.scheduleCertificateExpirationNotification(for: model, on: api)
                 return model
             }
     }
@@ -325,6 +327,16 @@ class AppCoordinator: RootViewCoordinator {
             #endif
         } else {
             showProfilesViewController(animated: animated)
+        }
+    }
+    
+    fileprivate func scheduleCertificateExpirationNotification(for certificate: CertificateModel, on api: Api) {
+        notificationsService.permissionGranted {
+            if $0 {
+                self.notificationsService.scheduleCertificateExpirationNotification(for: certificate, on: api)
+            } else {
+                os_log("Not Authorised", log: Log.general, type: .info)
+            }
         }
     }